Progress at the end of semester

+++ 3/samples/pow.c
@@ -2,10 +2,10 @@
 #include <math.h>
 
 int main(){
-  double a;
-  scanf("%lf", &a);
+  int a;
+  scanf("%d", &a);
   
   double b = pow(a, 2);//a^2 => a*a
   double c = pow(a, 2.31);//a^2 => ???
-  printf(" %lf %lf %lf\n", a, b, c);
+  printf(" %d %lf %lf\n", a, b, c);
 }
